+++ Assets/Scripts/Enemy/Enemy AI Scripts/ExplosiveEnemyAI.cs
@@ -69,9 +69,9 @@ public class ExplosiveEnemyAI : MonoBehaviour, IDamage
     }
 
     // Need to adjust this in all AI to work like lecture
-    public void takeDamage(int damage)
+    public void takeDamage(float damage)
     {
-        HP -= damage;
+        HP -= (int)damage;
         animator.SetTrigger("Hit1");
         StartCoroutine(flashDamage());
         facePlayer();
