+++ Assets/Scripts/Enemy/Enemy AI Scripts/ProjectileEnemyAI.cs
@@ -68,9 +68,9 @@ using UnityEngine.AI;
         }
     }
     // Need to adjust this in all AI to work like lecture
-    public void takeDamage(int damage)
+    public void takeDamage(float damage)
     {
-        HP -= damage;
+        HP -= (int)damage;
         animator.SetTrigger("Hit1");
         StartCoroutine(flashDamage());
         facePlayer();
