+++ Assets/Scripts/Enemy/IDamage.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamage 
 {
-    void takeDamage(int damage);
+    void takeDamage(float damage);
 }
